// apps/web/src/api/generated/captain-registrations/captainRegistrationsControllerListRegistrationsSortBy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type CaptainRegistrationsControllerListRegistrationsSortBy = typeof CaptainRegistrationsControllerListRegistrationsSortBy[keyof typeof CaptainRegistrationsControllerListRegistrationsSortBy];


export const CaptainRegistrationsControllerListRegistrationsSortBy = {
  submittedAt: 'submittedAt',
  tournamentStartsAt: 'tournamentStartsAt',
} as const;